Refactored Artwork component and added Blog link to Navbar.
