+++ client/src/public/stylesheets/Blog.css
@@ -1,0 +1,60 @@
+@import url('https://fonts.googleapis.com/css2?family=Merienda+One&display=swap');
+
+.polaroidFrame {
+  padding: 30px 15px 15px 15px;
+  border: 1px solid gray;
+  width: 268px;
+  border-radius: 2px;
+  box-shadow: 5px 5px 5px black;
+  background-color: white;
+  filter: sepia(.75);
+  transition: 0.3s;
+}
+
+.polaroidFrame:hover {
+  filter: sepia(0);
+}
+
+.polaroidImage {
+  display:flex;
+  align-items:center;
+  justify-content:center;
+  background-color: black;
+  min-height: 300px;
+  position: relative;
+  overflow: hidden;
+  border: 1px solid gray;
+}
+
+.polaroidImage img{
+  display: block;
+  max-width: 100%;
+}
+
+.polaroidFrame .title,
+.polaroidFrame .description,
+.polaroidFrame .author {
+  overflow: hidden;
+  display: -webkit-box;
+  -webkit-line-clamp: 2;
+  -webkit-box-orient: vertical;
+}
+
+.polaroidImage::after {
+  box-shadow: inset 0px 0px 5px 5px rgba(0, 0, 0, 0.25);
+  content: "";
+  display: block;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  width: 100%;
+}
+
+.title {
+  font-size: 20px;
+  font-family: "Merienda One", cursive;
+  text-align: center;
+  color:  black;
+  margin: 20px;
+}
+
